refactor: prepare for `hooks::utils`
