+++ src/hooks/replace_token.rs
@@ -1,5 +1,6 @@
 use serde_json::json;
 
+use super::utils::ReplaceToken;
 use super::{
     wrap_bad_hook_cmd,
     ActionHook,
@@ -13,14 +14,6 @@ use super::{
 use crate::errors::AliError;
 
 const USAGE: &str = "<TOKEN> <VALUE> <TEMPLATE> [OUTPUT]";
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct ReplaceToken {
-    token: String,
-    value: String,
-    template: String,
-    output: String,
-}
 
 struct HookReplaceToken {
     rp: ReplaceToken,
@@ -187,21 +180,6 @@ impl ToString for ReplaceToken {
     }
 }
 
-impl ReplaceToken {
-    fn replace(&self, s: &str) -> Result<String, AliError> {
-        let token = &format!("{} {} {}", "{{", self.token, "}}");
-
-        if !s.contains(token) {
-            return Err(AliError::BadHookCmd(format!(
-                "template {} does not contains token \"{token}\"",
-                self.template
-            )));
-        }
-
-        Ok(s.replace(token, &self.value))
-    }
-}
-
 #[test]
 fn test_parse_replace_token() {
     use std::collections::HashMap;
@@ -273,54 +251,5 @@ fn test_parse_replace_token() {
     for (cmd, expected) in tests {
         let actual = HookReplaceToken::try_from(cmd).unwrap();
         assert_eq!(expected, actual.rp);
-    }
-}
-
-#[test]
-fn test_replace_token() {
-    use std::collections::HashMap;
-
-    let tests = HashMap::from([
-        (
-            ReplaceToken {
-                token: String::from("PORT"),
-                value: String::from("3322"),
-                template: String::from("dummy.conf"),
-                output: String::from("dummy.conf"),
-            },
-            ("{{ PORT }} foo bar {{PORT}}", "3322 foo bar {{PORT}}"),
-        ),
-        (
-            ReplaceToken {
-                token: String::from("foo"),
-                value: String::from("bar"),
-                template: String::from("dummy.conf"),
-                output: String::from("dummy.conf"),
-            },
-            (
-                "{{ bar }} {{ foo }} {{ bar }} foo <{{ foo }}>",
-                "{{ bar }} bar {{ bar }} foo <bar>",
-            ),
-        ),
-        (
-            ReplaceToken {
-                token: String::from("foo"),
-                value: String::from("bar"),
-                template: String::from("dummy.conf"),
-                output: String::from("dummy.conf"),
-            },
-            (
-                "{ foo } {{ foo }} {{ foo }_} foo bar {{{ foo }}} {{ foo {{ foo }}}}",
-                "{ foo } bar {{ foo }_} foo bar {bar} {{ foo bar}}",
-            ),
-        ),
-    ]);
-
-    for (replace, (template, expected)) in tests {
-        let actual = replace
-            .replace(template)
-            .expect("failed to replace template {template}");
-
-        assert_eq!(expected, actual);
     }
 }
